selftests/bpf: fix use-after-free test on 32-bit
An update to the 'global_map_resize' test adds an 'extern int' which is
referenced by a global variable to avoid elision. However, the global is a
'long' type and so fails the skeleton's compile-time assertion when used
for 32-bit userspace. Fix by using a global 'int' instead.

Fixes: fa6f092cc0a0 ("libbpf: Fix possible use-after-free for externs")

+++ tools/testing/selftests/bpf/progs/test_global_map_resize.c
@@ -40,7 +40,7 @@ int percpu_arr[1] SEC(".data.percpu_arr");
  * but by having externs for the resizing tests it will cover this path.
  */
 extern int LINUX_KERNEL_VERSION __kconfig;
-long version_sink;
+int version_sink;
 
 SEC("tp/syscalls/sys_enter_getpid")
 int bss_array_sum(void *ctx)
